feat(driver): Move the service registry into the driver (#620)
