+++ build-extensions/build.gradle.kts
@@ -8,5 +8,5 @@ repositories {
 
 dependencies {
   implementation("net.kyori", "indra-common", "2.1.1")
-  implementation("com.google.code.gson", "gson", "2.8.9")
+  implementation("com.google.code.gson", "gson", "2.9.0")
 }
