+++ modules/bridge/src/main/java/eu/cloudnetservice/modules/bridge/BridgeManagement.java
@@ -16,8 +16,8 @@
 
 package eu.cloudnetservice.modules.bridge;
 
-import eu.cloudnetservice.cloudnet.common.registry.ServicesRegistry;
 import eu.cloudnetservice.cloudnet.driver.network.rpc.annotation.RPCValidation;
+import eu.cloudnetservice.cloudnet.driver.registry.ServicesRegistry;
 import eu.cloudnetservice.modules.bridge.config.BridgeConfiguration;
 import eu.cloudnetservice.modules.bridge.player.PlayerManager;
 import lombok.NonNull;
