+++ modules/bridge/src/main/java/eu/cloudnetservice/modules/bridge/platform/sponge/SpongeBridgeManagement.java
@@ -16,7 +16,7 @@
 
 package eu.cloudnetservice.modules.bridge.platform.sponge;
 
-import eu.cloudnetservice.cloudnet.common.registry.ServicesRegistry;
+import eu.cloudnetservice.cloudnet.driver.registry.ServicesRegistry;
 import eu.cloudnetservice.cloudnet.driver.service.ServiceInfoSnapshot;
 import eu.cloudnetservice.cloudnet.wrapper.Wrapper;
 import eu.cloudnetservice.modules.bridge.BridgeServiceHelper;
