+++ modules/signs/src/main/java/eu/cloudnetservice/modules/signs/platform/bukkit/BukkitSignsPlugin.java
@@ -17,7 +17,9 @@
 package eu.cloudnetservice.modules.signs.platform.bukkit;
 
 import eu.cloudnetservice.cloudnet.driver.CloudNetDriver;
+import eu.cloudnetservice.cloudnet.driver.registry.ServicesRegistry;
 import eu.cloudnetservice.modules.signs.GlobalChannelMessageListener;
+import eu.cloudnetservice.modules.signs.SignManagement;
 import eu.cloudnetservice.modules.signs.platform.PlatformSignManagement;
 import eu.cloudnetservice.modules.signs.platform.SignsPlatformListener;
 import eu.cloudnetservice.modules.signs.platform.bukkit.functionality.CommandSigns;
@@ -49,7 +51,7 @@ public class BukkitSignsPlugin extends JavaPlugin {
 
   @Override
   public void onDisable() {
-    BukkitSignManagement.defaultInstance().unregisterFromServiceRegistry();
+    ServicesRegistry.first(SignManagement.class).unregisterFromServiceRegistry();
     CloudNetDriver.instance().eventManager().unregisterListeners(this.getClass().getClassLoader());
   }
 }
