+++ modules/syncproxy/src/main/java/eu/cloudnetservice/modules/syncproxy/platform/bungee/BungeeCordSyncProxyManagement.java
@@ -16,7 +16,7 @@
 
 package eu.cloudnetservice.modules.syncproxy.platform.bungee;
 
-import eu.cloudnetservice.cloudnet.common.registry.ServicesRegistry;
+import eu.cloudnetservice.cloudnet.driver.registry.ServicesRegistry;
 import eu.cloudnetservice.modules.syncproxy.platform.PlatformSyncProxyManagement;
 import java.util.Collection;
 import java.util.UUID;
